Ignore times when comparing dates
Fixes #36462, but likely breaks rather a lot of apps, so is really just an investigation on my part. This PR changes date comparisons to ignore the time (which I think is correct), but will break all code that assumes that a date is at the exact time 00:00:00, and date + 00:00:01 is not equal to, but is greater than, the point in time date.

Currently Date to Time object comparison is handled in Rails using [this code](https://github.com/rails/rails/blob/master/activesupport/lib/active_support/core_ext/date/calculations.rb#L1360):

```ruby
  # Allow Date to be compared with Time by converting to DateTime and relying on the <=> from there.
  def compare_with_coercion(other)
    if other.is_a?(Time)
      to_datetime <=> other
    else
      compare_without_coercion(other)
    end
  end
  alias_method :compare_without_coercion, :<=>
  alias_method :<=>, :compare_with_coercion
```

This code converts an instance of Date to a DateTime object using `to_datetime`, then compares that to the Time object. This has some issues though, that I will attempt to work through.

In Ruby the Date object doesn't contain any time or zone data:

```
> Time.now
=> 2019-06-29 12:43:20 +0100
> Date.today.strftime("%z")
=> "+0000"
```

This means that calling to_datetime on it has no time or zone data to work with in the Date object itself.

```
$ irb -r 'active_support/all'
> td = Time.now.to_date
=> Sat, 29 Jun 2019
> td.strftime("%z")
=> "+0000"

> td.to_datetime
=> Sat, 29 Jun 2019 00:00:00 +0000
> td.to_datetime.strftime("%z")
=> "+0000"

> td.to_time
=> 2019-06-29 00:00:00 +0100
> td.to_time.strftime("%z")
=> "+0100"
```

And a new IRB session:

```
$ TZ=US/Eastern irb -r 'active_support/all'
> t = Time.now
=> 2019-06-29 07:51:56 -0400
> td = t.to_date
=> Sat, 29 Jun 2019
> td.to_time
=> 2019-06-29 00:00:00 -0400
> td.to_time.strftime("%z")
=> "-0400"
```

This difference between to_time and to_datetime seems like a bug in Ruby, though perhaps a hard one to change given the amount of code out there that will be relying on the current behaviour :)

Anyway, what this means for Rails is that the conversion of Date to DateTime creates *implicit* 00:00:00 +0000 time and zone data, which will influence the comparison between most dates and times:

```
$ TZ=US/Eastern irb -r 'active_support/all'
> current_time = Time.parse('Tue, 11 Jun 2019 00:30:00 EST -04:00')
=> 2019-06-11 00:30:00 -0500
> current_date = current_time.to_date
=> Tue, 11 Jun 2019
> current_date >= current_time
=> false
> current_date
=> Tue, 11 Jun 2019
> current_date.strftime("%z")
=> "+0000"
> current_time
=> 2019-06-11 00:30:00 -0500
```

Wether this is a bug I think is dependent on the behavior we want in the comparison. I would have thought that when comparing a date with a time we really are only interested in the date parts of the calculation, otherwise there's not much point comparing the two. In my mind comparing a date with a time is like saying "Is this the same day". If that is the case, dropping the time and zone information from the Time object allows the two objects to be compared correctly.

However, if we want dates to be compared to times based on the 00:00:00 hour of the date, then this isn't a bug.